refactor(utils): merge internal/pathutil into internal/utils
Consolidates the two utility packages into a single canonical
internal/utils package.

- Move all pathutil symbols (CheckDirectoryWritable,
  CheckFileDirectoryWritable, JoinAbsPath, NormalizeLibraryPath,
  RemoveEmptyDirs) into internal/utils/path.go
- Port pathutil tests to internal/utils/path_test.go
- Move config-dependent library.go out of utils into a new
  internal/library package to avoid the import cycle that would
  result from config importing utils
- Update all 7 importers to use utils instead of pathutil
- Delete internal/pathutil directory
