+++ internal/library/library.go
@@ -1,4 +1,4 @@
-package utils
+package library
 
 import (
 	"context"
